Add GameTimerContainer.tsx file.
